requirements.txt
opencv-python
mediapipe
numpy
pandas
PyYAML
tqdm